Create a component called "Next Button"
In the new component that is located at booking screens you can validate
