+++ src/components/index.js
@@ -7,3 +7,4 @@ export {My_flightlist}   from './flightlist/My_flightlist';
 export {PassengerPicker} from './Passengers/Passengers'
 export {LayoutRegister} from './Layout/index'
 export {BookingLayout} from './Layout/BookingLayout'
+export {NextButton} from './NextButton/NextButton'
