+++ src/screens/booking/Select_date.js
@@ -1,12 +1,12 @@
 import React, {useState} from 'react';
-import {View, Text, TouchableOpacity} from 'react-native';
-import {useNavigation} from '@react-navigation/native';
+import {View, Text} from 'react-native';
 import CalendarPicker from 'react-native-calendar-picker';
 import Style from './BookingStyles';
-import {BookingLayout} from '../../components/Layout/BookingLayout';
+import {BookingLayout, NextButton} from '../../components/index';
 
 export const Select_date = () => {
-  const navigation = useNavigation();
+  const [date, setDate] = useState(0);
+  const fullDate = date.toString()
 
   return (
     <BookingLayout
@@ -18,15 +18,9 @@ export const Select_date = () => {
       <View style={Style.to_container}>
         <Text style={Style.to}>Select Date</Text>
       </View>
-      <CalendarPicker />
-
-      <TouchableOpacity
-        style={Style.container_next}
-        onPress={() => {
-          navigation.navigate('Passenger');
-        }}>
-        <Text style={Style.text_next}>Next</Text>
-      </TouchableOpacity>
+      <CalendarPicker onDateChange={(value) => setDate (value)}/>
+      
+      <NextButton nextPage={'Passenger'} value={fullDate} />
     </BookingLayout>
   );
 };
