+++ src/screens/booking/To.js
@@ -1,30 +1,15 @@
 import React, {useState} from 'react';
 import {View, Text, TextInput, TouchableOpacity} from 'react-native';
 import Style from './BookingStyles';
-import {useNavigation} from '@react-navigation/native';
-import {BookingLayout} from '../../components/Layout/BookingLayout';
-export const To = () => {
-  const navigation = useNavigation();
+import {BookingLayout, NextButton} from '../../components/index';
+export const To = (props) => {
+  const {
+    route: {params: sumary},
+  } = props;
+
   const [textInputFrom, setTextInputFrom] = useState('');
-  let sumary = {
-    capitalOrigin: '',
-    countryOrigin: '',
-    capitalDestiny: '',
-    countryDestiny: '',
-    date: {
-      day: 0,
-      month: 0,
-      year: 0,
-    },
-    passengers: 0,
-  };
-  const checkTextInput = () => {
-    if (!textInputFrom.trim()) {
-      alert('Please Enter a City');
-      return;
-    }
-    navigation.navigate('Select_date', sumary);
-  };
+
+  
   return (
     <BookingLayout
       page={'From'}
@@ -42,9 +27,7 @@ export const To = () => {
         />
       </View>
 
-      <TouchableOpacity style={Style.container_next} onPress={checkTextInput}>
-        <Text style={Style.text_next}>Next</Text>
-      </TouchableOpacity>
-    </BookingLayout>
+      <NextButton value={textInputFrom} nextPage={'Select_date'}/>    
+      </BookingLayout>
   );
 };
